Separated authentication functions from App.js
